% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2_scales.R
\name{scale_color_glu}
\alias{scale_color_glu}
\title{Scale color Glu}
\usage{
scale_color_glu()
}
\description{
Discrete color scale using Glu color palette
}
\examples{
ggplot(...) + scale_color_glu
}
\seealso{
Other ggplot2_scales: \code{\link{scale_fill_glu}}
}
\author{
Beau Lucas
}
