% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2_scales.R
\name{scale_fill_glu}
\alias{scale_fill_glu}
\title{Scale fill Glu}
\usage{
scale_fill_glu()
}
\description{
Discrete fill scale using Glu color palette
}
\examples{
ggplot(...) + scale_fill_glu
}
\seealso{
Other ggplot2_scales: \code{\link{scale_color_glu}}
}
\author{
Beau Lucas
}
